Create a 3D video with an animated GPX route

Create a 3D video with an animated GPX route

video_animation(gpx_table = NULL, elevdata_long = NULL,
  number_of_screens = 10, make_gif = TRUE, theta = -10, zoom = 0.7,
  ffmpeg_path = "C:/Programme_2/ffmpeg/bin/", output_file_loc,
  overwrite = TRUE)

Arguments

gpx_table
(data.frame) gpx table created by get_enriched_gpx_table

elevdata_long
(data.frame) elevation data created by get_elevdata_long

number_of_screens
(numeric) A guess of the number of screens/frames that shall be shown. An amount of images of the 3D map will be created that suits this. The more screens there are the more fluent the animation is. Each screen though takes about 20-30 seconds to render on a decent PC

make_gif
(logical) Whether the screens shall be rendered out as a gif file. Just works for less than 24 screens

theta
Rotation around z-axis.

zoom
Zoom factor.

ffmpeg_path
(character) to render more than 24 screens ffmpeg is needed to be in the path. Therefore please install it and provide the link to the path here. The ffmpeg already in path will not be overwritten

output_file_loc
(character) Where to store the output mp4 file.

overwrite
(logical) Whether to overwrite the output fie